Add files via upload
Major but half-finished update I wouldn't normally release in this state, but need it for demo purposes. Added RGB support, copy buttons, support for WCAG AA and AAA, color sampling canvas, and more. A few minor UI bugs and some comments that need to be cleaned up, but I'm on a deadline, sue me.